requirements.txt
pydantic>=2.0
pydantic-settings>=2.0
openai>=1.0
httpx>=0.24
beautifulsoup4>=4.12
python-dotenv>=1.0
sqlalchemy>=2.0
aiosqlite>=0.19
jinja2>=3.1
click>=8.1
markdown>=3.5
schedule>=1.2

# Dev
pytest>=7.4
pytest-asyncio>=0.21
pytest-mock>=3.11
